refactor(validation): split the lock resolver out of the push-lock gate
Both files crossed the 500-line taste ceiling as the shell parsing grew. The
ratchet says to lower the count rather than raise the baseline, and there was
a real seam to cut on: the gate knows about Markdown units, the tracked-file
inventory and the CLI; resolving which paths a block of shell actually opens
is a separate question that knows nothing about fences or files.

`push_lock_resolver.py` now owns that question and `check_push_lock_paths.py`
consumes it, mirroring how the tests already split. The shell-parsing tests
move to their own module for the same reason. No behavior change: the same 68
tests pass, the corpus stays at 0 violations across 3520 files, and the public
entry points `scan_text`, `validate_push_lock_paths` and `main` are unmoved,
so `pre_pr_sequence` imports the same name from the same module.

Refs #4635
