+++ .serena/memories/memory-index.md
@@ -55,7 +55,7 @@
 |portability single traversal scan_all extra scan roots files_by_root required vendor-portability marker path drift declaration fenced code: [validation/validation-portability-scan-contract](validation/validation-portability-scan-contract.md) (753)
 |portability descendant symlink file directory escape repo root: [validation/descendant-symlink-portability-containment](validation/descendant-symlink-portability-containment.md) (210)
 |detector widening corpus probe measure before design count newly flagged gate false positive: [validation/measure-a-detector-widening-before-designing-it](validation/measure-a-detector-widening-before-designing-it.md) (409)
-|shared resolver latent defect inherited second caller granularity shell comment variable binding order: [validation/reusing-a-resolver-inherits-its-latent-defects](validation/reusing-a-resolver-inherits-its-latent-defects.md) (419)
+|shared resolver latent defect inherited second caller granularity shell comment variable binding order: [validation/reusing-a-resolver-inherits-its-latent-defects](validation/reusing-a-resolver-inherits-its-latent-defects.md) (460)
 |eval harness enumerate scripts naming underscore hyphen glob: [eval-harness-surface-enumeration](eval-harness-surface-enumeration.md) (681)
 |powershell ps1 psm1 module pester test discovery isolation: [skills-powershell-index](skills-powershell-index.md) (443), [skills-pester-testing-index](skills-pester-testing-index.md) (181), [powershell/pester-variable-scoping](powershell/pester-variable-scoping.md) (497), [powershell/powershell-variable-shadowing-detection](powershell/powershell-variable-shadowing-detection.md) (665), [patterns/pattern-thin-workflows](patterns/pattern-thin-workflows.md) (1373)
 |bash exit code cross-language contract AUTOFIX pre-commit hook: [skills-bash-integration-index](skills-bash-integration-index.md) (110)
